cleanup: reorganize group and board functions
