+++ source/client/connectGroups.ts
@@ -1,114 +1,4 @@
-type Orientation = 'vertical' | 'horizontal';
-
-interface Card {
-  id: string;
-  lowNum: string;
-  uprightFor: Orientation;
-}
-
-export interface Group {
-  number: number; // incremental number to keep track of the groups
-  cards: {
-    [x: number]: (null | Card)[];
-  };
-}
-
-function getNewGroupNumber() {
-  // TODO: actual implementation
-  return 5;
-}
-
-// TODO: calculation for more than one card
-export function createNewGroup(
-  orientation: Orientation,
-  ...cards: Card[]
-): Group {
-  const group: Group = {
-    number: getNewGroupNumber(),
-    cards: {},
-  };
-  if (cards.length === 1) {
-    group.cards[0] = [cards[0]!];
-  } else {
-    // orientation and then calculate how to put the cards in order
-  }
-
-  return group;
-}
-
-export function getConnectingCardNums(cardNum: string) {
-  // start with the number
-  let belowNum = String(parseInt(cardNum) - 1);
-  if (belowNum.length < 2) belowNum = '0' + belowNum;
-  let aboveNum = String(parseInt(cardNum) + 1);
-  if (aboveNum.length < 2) aboveNum = '0' + aboveNum;
-
-  return [belowNum, aboveNum];
-}
-
-interface Connection {
-  card: Card;
-  x: number;
-  y: number;
-}
-
-interface ConnectingCardLocation {
-  group: Group;
-  card: Card;
-  connection: Connection;
-  orientation: Orientation;
-}
-
-export function connectCardToGroup({
-  group,
-  card,
-  connection,
-  orientation,
-}: ConnectingCardLocation) {
-  // if the number of the card we are adding is greater than the card we are connecting to, it will go at the end of the row or column
-  // TODO: needs to actually calculate based on uprightFor value
-  const connectAtEnd = parseInt(card.lowNum) > parseInt(connection.card.lowNum);
-  const numCols = Object.keys(group.cards).length;
-
-  if (group.cards[connection.y]?.[connection.x]?.id !== connection.card.id) {
-    throw new Error('The connecting card details are not correct');
-  }
-
-  if (orientation === 'vertical') {
-    for (let i = 0; i < numCols; i++) {
-      const itemToAdd = i === connection.y ? card : null;
-      if (connectAtEnd) {
-        group.cards[i]?.splice(connection.x + 1, 0, itemToAdd);
-      } else {
-        group.cards[i]?.splice(connection.x, 0, itemToAdd);
-      }
-    }
-    return;
-  }
-
-  const numRows = group.cards[0]!.length;
-
-  if (connectAtEnd) {
-    group.cards[numCols] = [];
-    for (let i = 0; i < numRows; i++) {
-      const itemToAdd = i === connection.x ? card : null;
-      group.cards[numCols].push(itemToAdd);
-    }
-    return;
-  }
-  // if adding at the beginning of the row, we need to shift all the other columns to the right
-  for (let i = numCols; i >= 0; i--) {
-    if (i === 0) {
-      group.cards[0] = [];
-      for (let i = 0; i < numRows; i++) {
-        const itemToAdd = i === connection.x ? card : null;
-        group.cards[0].push(itemToAdd);
-      }
-      continue;
-    }
-    group.cards[i] = group.cards[i - 1]!;
-  }
-}
+import { Group, Connection, Orientation, Card } from './connectCardToGroup';
 
 interface ConnectGroupsArgs {
   smallerGroup: {
@@ -227,47 +117,4 @@ export function connectGroups({
     number: newGroupNum,
     cards: newCards,
   };
-}
-
-interface BoardCard {
-  id: string;
-  lowNum: string;
-  lowUprightForV: boolean;
-}
-
-export type GroupUI = (BoardCard | null)[][];
-
-export interface BoardUI {
-  [groupNum: string]: GroupUI;
-}
-
-export function genGroupUI(group: Group) {
-  const numCols = Object.keys(group.cards).length;
-  const numRows = group.cards[0]?.length;
-  if (!numRows) {
-    throw new Error(`somehow there are no rows in group: ${group.number}`);
-  }
-  const boardSpaces: GroupUI = [];
-
-  for (let x = 0; x < numCols; x++) {
-    boardSpaces[x] = [];
-    for (let y = 0; y < numRows; y++) {
-      const card = group.cards[x]?.[y];
-      if (card === null) {
-        boardSpaces[x]![y] = null;
-      } else {
-        if (!card) {
-          throw new Error(
-            `card was missing while creating the group: ${card}, at (${x},${y})`
-          );
-        }
-        boardSpaces[x]![y] = {
-          id: card.id,
-          lowNum: card.lowNum,
-          lowUprightForV: card.uprightFor === 'vertical',
-        };
-      }
-    }
-  }
-  return boardSpaces;
 }
